Add files via upload
Hello everyone,
for my final paper I have to work with CoolProp in Python. I should use texatherm 32 as the fluid. On line 39 it says I use water, but so this should be Texatherm 32. Texatherm 22 is in there but so I need texatherm 32....
I don't know how to add or create it. 
Thanks in advance